simulations/requirements.txt
paho-mqtt>=1.6.1
pyyaml>=6.0
pygame>=2.4.0
